requirements.txt
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.0.0
faiss-cpu>=1.7.4
langchain-text-splitters>=0.2.0
nltk>=3.8.1
spacy>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
tiktoken>=0.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
tqdm>=4.66.0
PyPDF2>=3.0.0
markdownify>=0.12.0
scikit-posthocs>=0.9.0
tabulate>=0.9.0
rouge-score>=0.1.2
